requirements.txt
aiogram==3.13.1
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
shortuuid==1.0.13
python-dotenv==1.0.1
asyncpg==0.29.0  # для PostgreSQL на Render